Story Security added

+++ server/app/controllers/blogTagController.js
@@ -1,51 +1,78 @@
 const BlogTag = require('../models/blogTag');
+const Blog = require('../models/blog');
 
 exports.addBlogTag = (req, res) => {
   const tag = req.body.tag;
   const postId = req.params.postId;
+  const bloggerId = req.user.id;
 
-  const tagObj = {
-    tag,
-    blog_id: postId
-  }
-  BlogTag.create(tagObj).then(blogTag => {
-    BlogTag.findAll({
-      where: {
+  Blog.findOne({
+    where: {
+      id: postId,
+      blogger_id: bloggerId
+    }
+  }).then(blog => {
+    if(blog){
+      const tagObj = {
+        tag,
         blog_id: postId
       }
-    }).then(blogTags => {
-      res.json(blogTags);
-    });
+      BlogTag.create(tagObj).then(blogTag => {
+        BlogTag.findAll({
+          where: {
+            blog_id: postId
+          }
+        }).then(blogTags => {
+          res.json(blogTags);
+        });
+      });
+    }
   });
 }
 
 exports.removeBlogTag = (req, res) => {
   const tagId = req.body.id;
   const postId = req.params.postId;
+  const bloggerId = req.user.id;
 
-  BlogTag.destroy({
+  Blog.findOne({
     where: {
-      id: tagId
+      id: postId,
+      blogger_id: bloggerId
     }
-  }).then(() => {
-    BlogTag.findAll({
-      where: {
-        blog_id: postId
-      }
-    }).then(blogTags => {
-      res.json(blogTags);
-    });
+  }).then(blog => {
+    if(blog){
+      BlogTag.destroy({
+        where: {
+          id: tagId
+        }
+      }).then(() => {
+        BlogTag.findAll({
+          where: {
+            blog_id: postId
+          }
+        }).then(blogTags => {
+          res.json(blogTags);
+        });
+      });
+    }
   });
 }
 
 exports.fetchBlogTag = (req, res) => {
-  const postId = req.params.postId;
-
+  const blogId = req.params.postId;
+  const bloggerId = req.user.id;
   BlogTag.findAll({
-    where: {
-      blog_id: postId
-    }
+    attributes: ['tag', 'id'],
+    include: [{
+      model: Blog,
+      attributes: [],
+      where: {
+        id: blogId,
+        blogger_id: bloggerId
+      }
+    }]
   }).then(blogTags => {
-    res.json(blogTags);
+    res.send(blogTags);
   });
 }
